feat: rework mobile navigation bar
