+++ front/src/modules/navigation/components/DesktopNavigationDrawer.tsx
@@ -1,0 +1,46 @@
+import { useCurrentUserTaskCount } from '@/activities/tasks/hooks/useCurrentUserDueTaskCount';
+import { useCommandMenu } from '@/command-menu/hooks/useCommandMenu';
+import { Favorites } from '@/favorites/components/Favorites';
+import { SettingsNavbar } from '@/settings/components/SettingsNavbar';
+import {
+  IconBell,
+  IconCheckbox,
+  IconSearch,
+  IconSettings,
+} from '@/ui/display/icon/index';
+import { useIsSettingsPage } from '../hooks/useIsSettingsPage';
+import MainNavbar from '@/ui/navigation/navigation-drawer/components/MainNavbar';
+import NavItem from '@/ui/navigation/navigation-drawer/components/NavItem';
+import { WorkspaceNavItems } from './WorkspaceNavItems';
+import { useIsTasksPage } from '@/navigation/hooks/useIsTasksPage';
+
+export const DesktopNavigationDrawer = () => {
+  const { toggleCommandMenu } = useCommandMenu();
+  const isSettingsPage = useIsSettingsPage();
+  const isTasksPage = useIsTasksPage()
+  const { currentUserDueTaskCount } = useCurrentUserTaskCount();
+
+  return isSettingsPage ? (
+    <SettingsNavbar />
+  ) : (
+    <MainNavbar>
+      <NavItem
+        label="Search"
+        Icon={IconSearch}
+        onClick={toggleCommandMenu}
+        keyboard={['⌘', 'K']}
+      />
+      <NavItem label="Notifications" to="/inbox" Icon={IconBell} soon />
+      <NavItem label="Settings" to="/settings/profile" Icon={IconSettings} />
+      <NavItem
+        label="Tasks"
+        to="/tasks"
+        active={isTasksPage}
+        Icon={IconCheckbox}
+        count={currentUserDueTaskCount}
+      />
+      <Favorites />
+      <WorkspaceNavItems />
+    </MainNavbar>
+  );
+};
