+++ front/src/modules/navigation/components/__stories__/MobileNavigationDrawer.stories.tsx
@@ -1,0 +1,23 @@
+import { Meta, StoryObj } from '@storybook/react';
+
+import { ComponentWithRouterDecorator } from '~/testing/decorators/ComponentWithRouterDecorator';
+
+import { MobileNavigationDrawer } from '../MobileNavigationDrawer';
+import { ComponentDecorator } from '~/testing/decorators/ComponentDecorator';
+import { SnackBarDecorator } from '~/testing/decorators/SnackBarDecorator';
+
+const meta: Meta<typeof MobileNavigationDrawer> = {
+  title: 'Modules/Navigation/MobileNavigationDrawer',
+  component: MobileNavigationDrawer,
+};
+
+export default meta;
+type Story = StoryObj<typeof MobileNavigationDrawer>;
+
+export const Default: Story = {
+  decorators: [
+    ComponentDecorator,
+    ComponentWithRouterDecorator,
+    SnackBarDecorator,
+  ],
+};
