+++ front/src/modules/navigation/hooks/useIsSettingsPage.ts
@@ -1,0 +1,4 @@
+import { useLocation } from 'react-router-dom';
+
+export const useIsSettingsPage = () =>
+  useLocation().pathname.match(/\/settings\//g) !== null;
