+++ front/src/modules/navigation/hooks/useIsTasksPage.ts
@@ -1,0 +1,4 @@
+import { useLocation } from 'react-router-dom';
+
+export const useIsTasksPage = () =>
+  useLocation().pathname === '/tasks';
