--- front/src/modules/ui/layout/hooks/useIsSubMenuNavbarDisplayed.ts
@@ -1,6 +1,0 @@
-import { useLocation } from 'react-router-dom';
-
-export const useIsSubMenuNavbarDisplayed = () => {
-  const currentPath = useLocation().pathname;
-  return currentPath.match(/\/settings\//g) !== null;
-};
