+++ front/src/modules/ui/layout/page/DefaultLayout.tsx
@@ -1,6 +1,5 @@
 import styled from '@emotion/styled';
 import { AnimatePresence, LayoutGroup } from 'framer-motion';
-import { useRecoilValue } from 'recoil';
 
 import { AuthModal } from '@/auth/components/Modal';
 import { useOnboardingStatus } from '@/auth/hooks/useOnboardingStatus';
@@ -8,25 +7,24 @@ import { OnboardingStatus } from '@/auth/utils/getOnboardingStatus';
 import { CommandMenu } from '@/command-menu/components/CommandMenu';
 import { KeyboardShortcutMenu } from '@/keyboard-shortcut-menu/components/KeyboardShortcutMenu';
 import { SignInBackgroundMockPage } from '@/sign-in-background-mock/components/SignInBackgroundMockPage';
-import { NavbarAnimatedContainer } from '@/ui/navigation/navbar/desktop-navbar/components/NavbarAnimatedContainer';
-import TabBar from '@/ui/navigation/navbar/mobile-navbar/tab-bar/TabBar';
-import { MOBILE_VIEWPORT } from '@/ui/theme/constants/theme';
-import { useIsMobile } from '@/ui/utilities/responsive/hooks/useIsMobile';
-import { DesktopNavbar } from '~/navbar/DesktopNavbar';
-import { MobileNavBar } from '~/navbar/MobileNavbar';
 
-import { isNavbarOpenedState } from '../states/isNavbarOpenedState';
+import { NavbarAnimatedContainer } from '@/ui/navigation/navigation-drawer/components/NavbarAnimatedContainer';
+import { useIsMobile } from '@/ui/utilities/responsive/hooks/useIsMobile';
+import { DesktopNavigationDrawer } from '@/navigation/components/DesktopNavigationDrawer';
+import { MobileNavigationDrawer } from '@/navigation/components/MobileNavigationDrawer';
+
+import { MobileNavigationBar } from '@/navigation/components/MobileNavigationBar';
+import { ReactNode } from 'react';
 
 const StyledLayout = styled.div`
   background: ${({ theme }) => theme.background.noisy};
   display: flex;
-  flex-direction: row;
+  flex-direction: column;
   height: 100vh;
   position: relative;
   scrollbar-color: ${({ theme }) => theme.border.color.medium};
-
   scrollbar-width: 4px;
-  width: 100vw;
+  width: 100%;
 
   *::-webkit-scrollbar {
     height: 4px;
@@ -43,46 +41,51 @@ const StyledLayout = styled.div`
   }
 `;
 
+const StyledPageContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+  flex: 1;
+`;
+
 const StyledMainContainer = styled.div`
   display: flex;
-  flex: 1;
+  flex: 0 1 100%;
   flex-direction: row;
   overflow: hidden;
-  @media (max-width: ${MOBILE_VIEWPORT}px) {
-    width: ${() => (useRecoilValue(isNavbarOpenedState) ? '0' : '100%')};
-  }
 `;
 
 type DefaultLayoutProps = {
-  children: React.ReactNode;
+  children: ReactNode;
 };
 
 export const DefaultLayout = ({ children }: DefaultLayoutProps) => {
   const onboardingStatus = useOnboardingStatus();
   const isMobile = useIsMobile();
-  
   return (
     <StyledLayout>
       <CommandMenu />
       <KeyboardShortcutMenu />
-      <NavbarAnimatedContainer>
-        {isMobile ? <MobileNavBar /> : <DesktopNavbar />}
-      </NavbarAnimatedContainer>
-      <StyledMainContainer>
-        {onboardingStatus && onboardingStatus !== OnboardingStatus.Completed ? (
-          <>
-            <SignInBackgroundMockPage />
-            <AnimatePresence mode="wait">
-              <LayoutGroup>
-                <AuthModal>{children}</AuthModal>
-              </LayoutGroup>
-            </AnimatePresence>
-          </>
-        ) : (
-          <>{children}</>
-        )}
-      </StyledMainContainer>
-      <TabBar />
+      <StyledPageContainer>
+        <NavbarAnimatedContainer>
+          {isMobile ? <MobileNavigationDrawer /> : <DesktopNavigationDrawer />}
+        </NavbarAnimatedContainer>
+        <StyledMainContainer>
+          {onboardingStatus &&
+          onboardingStatus !== OnboardingStatus.Completed ? (
+            <>
+              <SignInBackgroundMockPage />
+              <AnimatePresence mode="wait">
+                <LayoutGroup>
+                  <AuthModal>{children}</AuthModal>
+                </LayoutGroup>
+              </AnimatePresence>
+            </>
+          ) : (
+            children
+          )}
+        </StyledMainContainer>
+      </StyledPageContainer>
+      {isMobile && <MobileNavigationBar />}
     </StyledLayout>
   );
 };
