+++ front/src/modules/ui/layout/page/PageContainer.tsx
@@ -1,15 +1,9 @@
 import styled from '@emotion/styled';
 
-type PageContainerProps = {
-  children: JSX.Element | JSX.Element[];
-};
-
 const StyledContainer = styled.div`
   display: flex;
   flex-direction: column;
   width: 100%;
 `;
 
-export const PageContainer = ({ children }: PageContainerProps) => (
-  <StyledContainer>{children}</StyledContainer>
-);
+export { StyledContainer as PageContainer };
