+++ front/src/modules/ui/layout/page/RightDrawerContainer.tsx
@@ -4,29 +4,28 @@ import { RightDrawer } from '@/ui/layout/right-drawer/components/RightDrawer';
 import { useIsMobile } from '@/ui/utilities/responsive/hooks/useIsMobile';
 
 import { PagePanel } from './PagePanel';
+import { MOBILE_VIEWPORT } from '@/ui/theme/constants/theme';
+import { ReactNode } from 'react';
 
 type RightDrawerContainerProps = {
-  children: JSX.Element | JSX.Element[];
-  topMargin?: number;
+  children: ReactNode;
 };
 
-const StyledMainContainer = styled.div<{
-  topMargin: number;
-  isMobile?: boolean;
-}>`
+const StyledMainContainer = styled.div`
   background: ${({ theme }) => theme.background.noisy};
   display: flex;
-
   flex-direction: row;
   gap: ${({ theme }) => theme.spacing(2)};
-  height: calc(
-    100% -
-      ${(props) => (props.isMobile ? props.topMargin + 44 : props.topMargin)}px
-  );
-
+  height: 100%;
   padding-bottom: ${({ theme }) => theme.spacing(3)};
   padding-right: ${({ theme }) => theme.spacing(3)};
-  width: calc(100% - ${({ theme }) => theme.spacing(3)});
+  width: 100%;
+  box-sizing: border-box;
+
+  @media (max-width: ${MOBILE_VIEWPORT}px) {
+    padding-left: ${({ theme }) => theme.spacing(3)};
+    padding-bottom: 0;
+  }
 `;
 
 type LeftContainerProps = {
@@ -42,15 +41,11 @@ const StyledLeftContainer = styled.div<LeftContainerProps>`
 
 export const RightDrawerContainer = ({
   children,
-  topMargin,
-}: RightDrawerContainerProps) => {
-  const isMobile = useIsMobile();
-  return (
-    <StyledMainContainer topMargin={topMargin ?? 0} isMobile={isMobile}>
-      <StyledLeftContainer>
-        <PagePanel>{children}</PagePanel>
-      </StyledLeftContainer>
-      <RightDrawer />
-    </StyledMainContainer>
-  );
-};
+}: RightDrawerContainerProps) => (
+  <StyledMainContainer>
+    <StyledLeftContainer>
+      <PagePanel>{children}</PagePanel>
+    </StyledLeftContainer>
+    <RightDrawer />
+  </StyledMainContainer>
+);
