+++ front/src/modules/ui/layout/states/isNavbarOpenedState.ts
@@ -1,10 +1,6 @@
 import { atom } from 'recoil';
 
-import { MOBILE_VIEWPORT } from '@/ui/theme/constants/theme';
-
-const isMobile = window.innerWidth <= MOBILE_VIEWPORT;
-
-export const isNavbarOpenedState = atom({
-  key: 'ui/isNavbarOpenedState',
-  default: !isMobile,
+export const navigationDrawerState = atom<'main' | 'settings' | ''>({
+  key: 'ui/navigationDrawerState',
+  default: '',
 });
