--- front/src/modules/ui/layout/states/isNavbarSwitchingSizeState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const isNavbarSwitchingSizeState = atom({
-  key: 'ui/isNavbarSwitchingSizeState',
-  default: true,
-});
