--- front/src/modules/ui/navigation/navbar/desktop-navbar/components/NavBackButton.tsx
@@ -1,56 +1,0 @@
-import { useNavigate } from 'react-router-dom';
-import styled from '@emotion/styled';
-import { useRecoilState } from 'recoil';
-
-import { IconChevronLeft } from '@/ui/display/icon/index';
-import { isNavbarSwitchingSizeState } from '@/ui/layout/states/isNavbarSwitchingSizeState';
-
-type NavBackButtonProps = {
-  title: string;
-};
-
-const StyledIconAndButtonContainer = styled.button`
-  align-items: center;
-  background: inherit;
-  border: none;
-  color: ${({ theme }) => theme.font.color.secondary};
-  cursor: pointer;
-  display: flex;
-  flex-direction: row;
-  font-size: ${({ theme }) => theme.font.size.lg};
-  font-weight: ${({ theme }) => theme.font.weight.semiBold};
-  gap: ${({ theme }) => theme.spacing(1)};
-  padding: ${({ theme }) => theme.spacing(1)};
-  width: 100%;
-`;
-
-const StyledContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-`;
-
-const NavBackButton = ({ title }: NavBackButtonProps) => {
-  const navigate = useNavigate();
-  const [, setIsNavbarSwitchingSize] = useRecoilState(
-    isNavbarSwitchingSizeState,
-  );
-
-  return (
-    <>
-      <StyledContainer>
-        <StyledIconAndButtonContainer
-          onClick={() => {
-            setIsNavbarSwitchingSize(true);
-            navigate('/', { replace: true });
-          }}
-        >
-          <IconChevronLeft />
-          <span>{title}</span>
-        </StyledIconAndButtonContainer>
-      </StyledContainer>
-    </>
-  );
-};
-
-export default NavBackButton;
