--- front/src/modules/ui/navigation/navbar/desktop-navbar/constants/index.ts
@@ -1,11 +1,0 @@
-export const leftNavbarWidth = {
-  mobile: 'calc(100% - 16px)',
-  desktop: '220px',
-};
-
-export const leftSubMenuNavbarWidth = {
-  mobile: 'calc(100% - 16px)',
-  desktop: '520px',
-};
-
-export const githubLink = 'https://github.com/twentyhq/twenty';
