+++ front/src/modules/ui/navigation/navigation-bar/components/__stories__/NavigationBar.stories.tsx
@@ -1,0 +1,33 @@
+import { Meta, StoryObj } from '@storybook/react';
+
+import { ComponentWithRouterDecorator } from '~/testing/decorators/ComponentWithRouterDecorator';
+
+import { NavigationBar } from '../NavigationBar';
+import { ComponentDecorator } from '~/testing/decorators/ComponentDecorator';
+import {
+  IconList,
+  IconSearch,
+  IconCheckbox,
+  IconSettings,
+} from '@/ui/display/icon';
+
+const meta: Meta<typeof NavigationBar> = {
+  title: 'UI/Navigation/NavigationBar/NavigationBar',
+  component: NavigationBar,
+  args: {
+    activeItemName: 'main',
+    items: [
+      { name: 'main', Icon: IconList, onClick: () => undefined },
+      { name: 'search', Icon: IconSearch, onClick: () => undefined },
+      { name: 'tasks', Icon: IconCheckbox, onClick: () => undefined },
+      { name: 'settings', Icon: IconSettings, onClick: () => undefined },
+    ],
+  },
+};
+
+export default meta;
+type Story = StoryObj<typeof NavigationBar>;
+
+export const Default: Story = {
+  decorators: [ComponentDecorator, ComponentWithRouterDecorator],
+};
