+++ front/src/modules/ui/navigation/navigation-drawer/components/MainNavbar.tsx
@@ -1,0 +1,45 @@
+import { useState } from 'react';
+import styled from '@emotion/styled';
+
+import NavItemsContainer from './NavItemsContainer';
+import NavWorkspaceButton from './NavWorkspaceButton';
+import SupportChat from './SupportChat';
+
+type MainNavbarProps = {
+  children: React.ReactNode;
+};
+
+const StyledContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  justify-content: space-between;
+  margin-bottom: ${({ theme }) => theme.spacing(2.5)};
+  width: 100%;
+  padding: ${({ theme }) => theme.spacing(2)};
+  box-sizing: border-box;
+`;
+
+const MainNavbar = ({ children }: MainNavbarProps) => {
+  const [isHovered, setIsHovered] = useState(false);
+
+  const handleHover = () => {
+    setIsHovered(true);
+  };
+
+  const handleMouseLeave = () => {
+    setIsHovered(false);
+  };
+
+  return (
+    <StyledContainer>
+      <div onMouseEnter={handleHover} onMouseLeave={handleMouseLeave}>
+        <NavWorkspaceButton showCollapseButton={isHovered} />
+        <NavItemsContainer>{children}</NavItemsContainer>
+      </div>
+      <SupportChat />
+    </StyledContainer>
+  );
+};
+
+export default MainNavbar;
