+++ front/src/modules/ui/navigation/navigation-drawer/components/NavItemsContainer.tsx
