+++ front/src/modules/ui/navigation/navigation-drawer/components/NavTitle.tsx
