+++ front/src/modules/ui/navigation/navigation-drawer/components/NavWorkspaceButton.tsx
