+++ front/src/modules/ui/navigation/navigation-drawer/constants/index.ts
@@ -1,0 +1,6 @@
+export const desktopNavDrawerWidths = {
+  menu: '236px',
+  submenu: '536px',
+};
+
+export const githubLink = 'https://github.com/twentyhq/twenty';
